requirements.txt
Flask==2.3.3
psycopg[binary]==3.2.9
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
